Add CentOS 6 config
Lots of weird voodoo here, because CentOS has such ancient software.

Red Hat provide a thing called "Software Collections" to get (for e.g.)
a newer Python. This then needs to be wrapped with 'scl enable python27'
so it can set PATH and LD_LIBRARY_PATH. This is kind of nasty, but
works.