FreeBSD: Remove login shell files and firstboot
Creating /firstboot means that on startup the system will run configinit
to install users and packages, but this isn't necessary. It will also
install OS updates which triggers a reboot, breaking Buildbot.

The login files were mostly an annoyance as FreeBSD runs fortune with
the freebsd-tips file on login by default; this is emitted on the
console as Buildbot is started.