Debian 9: Separate images for libssl 1.0 and 1.1
Stretch uses libssl 1.1 by default. Box currently doesn't compile with
1.1 (GitHub issue boxbackup/boxbackup#16).

libssl 1.0 is available as a separate package - replacing libssl-dev, so
this creates two AMIs, one for each libssl version.

As part of this we need to use separate usernames for Buildbot. Packer
allows specifying a name value to each builder, so we use that builder
name as the Buildbot name. This requires passing the PACKER_BUILD_NAME
environment variable around.