Add Debian 9 i386 config
The Debian package maintainer (Reinhard Tartler) pointed out that i386
tests are failing in the Debian build cluster.

EC2 only provides 64-bit images on current generation hardware, so in
order to provide an i386 environment we need to use a chroot.

This adds two new builders to the Debian 9 template and a utils.sh file
to manage setting up the same settings both in and out of the chroot.