Debian 9 i386: Bind mount /etc/resolv.conf
The chroot had a copy of resolv.conf from a different VPC, so DNS
resolution didn't work.

This changes the systemd service to bind mount resolv.conf when the
service starts. This is the most flexible way to ensure resolv.conf is
always up to date; having the OS do the bind mount from /etc/fstab
doesn't always work as cloud-init seems to change the inode of
/etc/fstab, providing a stale mount in the chroot.